+++ packages/client/hmi-client/src/components/workflow/ops/calibrate-ensemble-ciemss/mod.ts
@@ -1,6 +1,6 @@
 import { CalibrateEnsembleCiemssOperation as operation } from './calibrate-ensemble-ciemss-operation';
 import node from './tera-calibrate-ensemble-node-ciemss.vue';
-import drilldown from './tera-calibrate-ensemble-ciemss.vue';
+import drilldown from './tera-calibrate-ensemble-ciemss-drilldown.vue';
 
 const name = operation.name;
